Intel IOMMU: bugref:9967 Build fix.
svn:sync-xref-src-repo-rev: r145183

+++ src/VBox/Devices/Bus/DevIommuIntel.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevIommuIntel.cpp 89731 2021-06-16 06:07:12Z ramshankar.venkataraman@oracle.com $ */
+/* $Id: DevIommuIntel.cpp 89732 2021-06-16 06:21:45Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  * IOMMU - Input/Output Memory Management Unit - Intel implementation.
  */
@@ -53,7 +53,7 @@
 
 /** Checks if the MMIO offset is valid. */
 #define DMAR_IS_MMIO_OFF_VALID(a_off)               (   (a_off) < DMAR_MMIO_GROUP_0_OFF_END \
-                                                     || (a_off) - (uint16_t)DMAR_MMIO_GROUP_1_OFF_FIRST < DMAR_MMIO_GROUP_1_SIZE)
+                                                     || (a_off) - (uint16_t)DMAR_MMIO_GROUP_1_OFF_FIRST < (uint16_t)DMAR_MMIO_GROUP_1_SIZE)
 
 /** Acquires the DMAR lock but returns with the given busy error code on failure. */
 #define DMAR_LOCK_RET(a_pDevIns, a_pThisCC, a_rcBusy) \
